requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
faiss-cpu
openai